Complete GLM-5.2 PR #13 residual defect audit
Normalize all expert-router TopK indices and the indexer full-context fast path to signed int64. Document every split, TopK, and intentional int32 disposition across the live module, with regression coverage.

Provenance: Infinity-AI-Institute/trainium PR #13 (deepseek-v4-flash-base)

Callsign: codex-bravo

+++ test/unit/model/glm52_moe_dsa/test_indexer.py
@@ -182,5 +182,5 @@ def test_full_context_selection_bypasses_topk_and_preserves_all_positions() -> N
         position_ids=torch.tensor([[0], [0]]),
     )
 
-    expected = torch.arange(4, dtype=torch.int32).reshape(1, 1, 4).expand(2, -1, -1)
+    expected = torch.arange(4, dtype=torch.int64).reshape(1, 1, 4).expand(2, -1, -1)
     torch.testing.assert_close(indices, expected)
